Rename mixxx.{cpp,h} to mixxxmainwindow.{cpp,h}
