+++ src/preferences/dialog/dlgprefwaveform.cpp
@@ -2,7 +2,7 @@
 
 #include "library/dao/analysisdao.h"
 #include "library/library.h"
-#include "mixxx.h"
+#include "mixxxmainwindow.h"
 #include "moc_dlgprefwaveform.cpp"
 #include "preferences/waveformsettings.h"
 #include "util/db/dbconnectionpooled.h"
